% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{availableGeneRanges}
\alias{availableGeneRanges}
\title{Return an GRanges object containing all unique ranges for the given gene}
\usage{
availableGeneRanges(gene, build = "GRCh38")
}
\arguments{
\item{gene}{Gene name}

\item{build}{Genome build. One of "GRCh38" or "GRCh37"}
}
\value{
\code{GRanges} object with unique ranges for all haplotypes (star alleles)
for the desired gene
}
\description{
Return a GRanges object containing all unique ranges for a given PGx gene 
definition. This function collects all haplotype ranges for the given gene
and returns only ranges for unique positions across all definitions. This 
function is mostly intended for internal use in \code{readPGx()} where the
returned ranges are used to collect all positions annotated in the sample 
VCF.
}
\examples{
pgxGeneRanges("CYP2C19")
}
