% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{availableGenes}
\alias{availableGenes}
\title{Return all available PGx genes}
\usage{
availableGenes(build = c("GRCh38", "GRCh37"))
}
\arguments{
\item{build}{Genome build. One of "GRCh38" or "GRCh37"}
}
\value{
character vector of available gene names
}
\description{
Return a vector of all PGx genes for which an allele definition exists.
}
\examples{
pgxGenes()
}
