% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{availableHaplotypeRanges}
\alias{availableHaplotypeRanges}
\title{Return a VRanges object of the unique ranges for the given haplotype}
\usage{
availableHaplotypeRanges(haplotype, build = c("GRCh38", "GRCh37"))
}
\arguments{
\item{haplotype}{Haplotype (star allele) name}

\item{build}{Genome build. One of "GRCh38" or "GRCh37"}
}
\value{
\code{GRanges} object with unique ranges for the given haplotype (star allele)
}
\description{
Return a VRanges object for the all ranges in the haplotype definition for 
the given haplotype.
}
\examples{
pgxHaplotypeRanges("CYP2C19_2")
}
