% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{availableHaplotypes}
\alias{availableHaplotypes}
\title{Return all available PGx haplotypes (star alleles)}
\usage{
availableHaplotypes(build = "GRCh38")
}
\arguments{
\item{build}{Genome build. One of "GRCh38" or "GRCh37"}
}
\value{
character vector of available defined star alleles (haplotypes)
}
\description{
Return a vector of haplotype names (star alleles) for which an allele 
definition exists.
}
\examples{
pgxHaplotypes()
}
