% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{buildReferenceDataframe,PGx-method}
\alias{buildReferenceDataframe,PGx-method}
\title{Build the reference data frame from the callable allele positions}
\usage{
\S4method{buildReferenceDataframe}{PGx}(x)
}
\value{
data.frame containing haplotype definitions
}
\description{
The reference data.frame consists of rows for all positions present in the 
sample VCF and columns for every callable allele haplotype. Since not all 
positions are present in every haplotype, missing positions for each 
haplotype are filled in with reference bases.
}
