% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{callableAlleles<-,PGx-method}
\alias{callableAlleles<-,PGx-method}
\title{Setter method for callableAlleles slot of PGx object}
\usage{
\S4method{callableAlleles}{PGx}(x) <- value
}
\description{
Setter method for callableAlleles slot of PGx object
}
