% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{callableAlleles,PGx-method}
\alias{callableAlleles,PGx-method}
\title{Getter method for callableAlleles slot of a PGx object}
\usage{
\S4method{callableAlleles}{PGx}(x)
}
\value{
Character vector of alleles that can be called
}
\description{
The callableAlleles slot contains a character vector of allele names for all
allele definitions that are completely represented in the sample VCF and are
thus deemed 'callable'.
}
