% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{callDiplotypes}
\alias{callDiplotypes}
\title{Perform diplotype calling pipeline}
\usage{
callDiplotypes(vcf, gene, build = "GRCh38", phased = TRUE)
}
\arguments{
\item{vcf}{Path to VCF file}

\item{gene}{The PGx gene to perform allele calling for. Must be one of \code{availableGenes()}}

\item{build}{The genome build. One of "GRCh38" or "GRCh37".}

\item{phased}{Logical value indicating whether or not the input data are 
phased. TRUE is the only accepted value currently.}
}
\value{
DataFrame with sample names as rows and column as PGx gene with allele calls
}
\description{
This function is a wrapper around the main pipeline steps for calling
diplotypes from VCF files. It wraps the following functions: (1) \code{readPGx()}
(2) \code{determineCallableAlleles()} (3) \code{buildReferenceDataFrame()} (4)
\code{convertGTtoNucleotides()} (5) \code{callPhasedDiplotypes()} and produces
a single DataFrame with allele calls for all samples in the VCF as output.
}
