% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{callPhasedDiplotype,PGx-method}
\alias{callPhasedDiplotype,PGx-method}
\title{Call the phased diplotypes for all samples in the PGx object}
\usage{
\S4method{callPhasedDiplotype}{PGx}(x)
}
\value{
DataFrame with allele calls for every sample in PGx
}
\description{
Returns a DataFrame of allele calls for every sample in the single allele PGx
object.
}
