% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{callPhasedDiplotypes,PGx-method}
\alias{callPhasedDiplotypes,PGx-method}
\title{Call phased diplotypes for all samples and callable alleles of a PGx object}
\usage{
\S4method{callPhasedDiplotypes}{PGx}(x)
}
\value{
DataFrame containing allele calls for all samples of the PGx object
for each callable allele
}
\description{
Call phased diplotypes for all samples and callable alleles of a PGx object
}
