% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{checkCallable}
\alias{checkCallable}
\title{Check if all haplotype ranges are in the PGx object for the given gene}
\usage{
checkCallable(x, gr)
}
\arguments{
\item{x}{PGx object}

\item{gr}{GRanges object}
}
\value{
logical vector indicating if all positions are present in the
sample VCF for the given haplotype.
}
\description{
Determine callable positions by checking to see if all positions in the allele
definition are present in the VCF file.
}
