% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{checkComplete}
\alias{checkComplete}
\title{Check if all haplotype ranges are in the PGx object for the given gene}
\usage{
checkComplete(gr, x)
}
\arguments{
\item{gr}{GRanges object}

\item{x}{PGx object}
}
\value{
logical vector indicating if all positions are present in the
sample VCF for the given haplotype.
}
\description{
Check if all haplotype ranges are in the PGx object for the given gene
}
