% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-diplotype-matcher.R
\name{checkOverlaps}
\alias{checkOverlaps}
\title{Check PGx object for overlapping ranges}
\usage{
checkOverlaps(pgx, ref)
}
\arguments{
\item{pgx}{PGx object}

\item{ref}{VRanges object containing haplotype definitions}
}
\value{
boolean value indicating whether all defined positions are present in
the sample PGx object
}
\description{
Return a boolean indicating whether the allele definition ranges are 
completely represented in the sample VCF. i.e. all positions in the 
reference are also present in the sample VCF. This subset represents the 
callable' alleles
}
