% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{cleanJoined}
\alias{cleanJoined}
\title{Clean the merged data.table}
\usage{
cleanJoined(x, gene)
}
\arguments{
\item{x}{data.table after joining definition to genotype dt}

\item{gene}{Gene name used to determine which columns contain allele definitions}
}
\description{
Clean the merged data.table
}
