% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{convertGTtoNucleotides,PGx-method}
\alias{convertGTtoNucleotides,PGx-method}
\title{Convert genotype calls to nucleotides}
\usage{
\S4method{convertGTtoNucleotides}{PGx}(x)
}
\value{
PGx object with geno(x)$GT converted to nucleotide representation
}
\description{
Convert genotype calls to nucleotides
}
