% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-pgx.R
\name{createIDs}
\alias{createIDs}
\title{Construct row IDs for the PGx object}
\usage{
createIDs(x)
}
\arguments{
\item{x}{PGx object}
}
\value{
character vector of IDs
}
\description{
The \code{readPGx()} function does not read in the ID column of the VCF by
default so rownames must be constructed manually. This is intentional so
definition ids and VCF ids can match after conversion to a genotype matrix.
}
