% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{detailPhasedCalls,PGx-method}
\alias{detailPhasedCalls,PGx-method}
\title{Detail phased diplotype calls}
\usage{
\S4method{detailPhasedCalls}{PGx}(x)
}
\value{
List of lists of matrices, one element for each sample in the PGx 
object. Each element contains a position by star allele boolean matrix for 
each haplotype.
}
\description{
This method returns a position by star allele boolean matrix for each 
haplotype for all samples in the object indicating which positions exactly 
match the reference DataFrame.
}
