% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.callSamplePhasedDiplotype}
\alias{.callSamplePhasedDiplotype}
\title{Call the phased diplotype for a single sample (column) of the genotype matrix}
\usage{
.callSamplePhasedDiplotype(x, allele, REF, ALT)
}
\description{
Call the phased diplotype for a single sample (column) of the genotype matrix
}
