% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-pgx.R
\name{.checkSwapSeqLevels}
\alias{.checkSwapSeqLevels}
\title{Check VCF and reference seqLevels}
\usage{
.checkSwapSeqLevels(file, ref, build)
}
\description{
Attempt to swap the reference seqLevels to match the VCF seqLevels before
importing the data. This step is necessary since we use ranges as a param
argument to the readVcf function which requires the same seqLevels in both
the reference ranges and the input VCF to be identical.
}
