% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{.extractHaplotypeRanges}
\alias{.extractHaplotypeRanges}
\title{Extract a GRangesList for all defined haplotypes of the PGx object}
\usage{
.extractHaplotypeRanges(x)
}
\description{
Extract a GRangesList for all defined haplotypes of the PGx object
}
