% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{.geno2geno}
\alias{.geno2geno}
\title{Taken from:
https://github.com/Bioconductor/VariantAnnotation/blob/54d2de6c7a1aa76f404076db79077cd05114edf2/R/methods-readVcf.R}
\usage{
.geno2geno(lst, ALT = NULL, REF = NULL, GT = NULL)
}
\description{
Taken from:
https://github.com/Bioconductor/VariantAnnotation/blob/54d2de6c7a1aa76f404076db79077cd05114edf2/R/methods-readVcf.R
}
