% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{.getHaplotypes}
\alias{.getHaplotypes}
\title{Extract haplotype strings for a vector of genotype strings}
\usage{
.getHaplotypes(x, sep = "|")
}
\description{
Extract haplotype strings for a vector of genotype strings
}
