% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{.isCallable}
\alias{.isCallable}
\title{Determine if one set of ranges completely encompasses another}
\usage{
.isCallable(q, s)
}
\description{
Determine if one set of ranges completely encompasses another
}
