% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{extractHaplotypeRanges,PGx-method}
\alias{extractHaplotypeRanges,PGx-method}
\title{Extract a GRangesList for all defined haplotypes of the PGx object}
\usage{
\S4method{extractHaplotypeRanges}{PGx}(x)
}
\value{
GRangesList of haplotype definitions
}
\description{
Extract a GRangesList for all defined haplotypes of the PGx object
}
