% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{extractStar}
\alias{extractStar}
\title{Extract the star allele string from the call string}
\usage{
extractStar(x, gene, alleles)
}
\arguments{
\item{x}{string of called star alleles}

\item{gene}{Gene name to remove from the star allele}

\item{alleles}{Vector of alleles that were kept in the analysis}
}
\value{
character string of star allele calls for the given sample
}
\description{
If the position is empty from the call string --> no exact match then assume
reference allele and assign the *1 allele
}
