% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{genotypeMatrixToDT}
\alias{genotypeMatrixToDT}
\title{Extract genotype matrix as data.table and pivot longer}
\usage{
genotypeMatrixToDT(x)
}
\arguments{
\item{x}{PGx object}
}
\value{
data.table with rows for every sample x position + GT
}
\description{
Extract genotype matrix as data.table and pivot longer
}
