% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{genotypeToNucleotides}
\alias{genotypeToNucleotides}
\title{Convert genotype calls to nucleotides}
\usage{
genotypeToNucleotides(x)
}
\arguments{
\item{x}{data.table}
}
\value{
data.table
}
\description{
Convert genotype calls to nucleotides
}
