% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getCallableAlleles,PGx-method}
\alias{getCallableAlleles,PGx-method}
\title{Return a vector of allele names that able to be called for the given PGx object}
\usage{
\S4method{getCallableAlleles}{PGx}(x)
}
\value{
PGx object with callableAlleles slot filled
}
\description{
Callable alleles are alleles where all defined positions are present in the 
sample VCF. The function will also add the vector of the callable alleles to
the callableAlleles slot of the PGx object
}
