% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{getGenotypeDT}
\alias{getGenotypeDT}
\title{Extract the genotype matrix from the PGx object and melt longer}
\usage{
getGenotypeDT(x)
}
\arguments{
\item{x}{PGx object}
}
\value{
data.table
}
\description{
Extract the genotype matrix from the PGx object and melt longer
}
