% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-diplotype-matcher.R
\name{getHaplotypeRanges}
\alias{getHaplotypeRanges}
\title{Return all haplotype ranges for the given PGx object}
\usage{
getHaplotypeRanges(pgx)
}
\arguments{
\item{pgx}{PGx object}
}
\value{
list of VRanges
}
\description{
Return all haplotype ranges for the given PGx object
}
