% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{hammingDistance}
\alias{hammingDistance}
\title{Hamming distance between vectors}
\usage{
hammingDistance(x, ref, cols)
}
\arguments{
\item{x}{data.table}

\item{ref}{reference column of the data.table}

\item{cols}{vector of column names containing haplotype definitions}
}
\description{
Hamming distance between vectors
}
