% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyrius.R
\name{installCyrius}
\alias{installCyrius}
\title{Install Cyrius Python Dependencies}
\usage{
installCyrius(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a 
method that will work in the local environment. Change the default to force 
a specific installation method. Note that the "virtualenv" method is not 
available on Windows.}

\item{conda}{The path to a conda executable. Use "auto" to allow reticulate 
to automatically find an appropriate conda binary. See Finding Conda and 
\link{reticulate::conda_binary()} for more details.}
}
\description{
This function will download and install the Python dependencies needed to 
run Cyrius from within R.
}
