% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{joinSampleToDef}
\alias{joinSampleToDef}
\title{Merge the sample data.tables to the definitions}
\usage{
joinSampleToDef(x, y)
}
\arguments{
\item{x}{sample data.table}

\item{y}{haplotype definition data.table}
}
\description{
Merge the sample data.tables to the definitions
}
