% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pgxBuild,PGx-method}
\alias{pgxBuild,PGx-method}
\title{Getter method for the pgxBuild slot of a PGx object}
\usage{
\S4method{pgxBuild}{PGx}(x)
}
\value{
Name of the genome build
}
\description{
The pgxBuild slot contains the build information for which the PGx object
is instantiated with.
}
