% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{pgxCallableAlleles<-,PGx-method}
\alias{pgxCallableAlleles<-,PGx-method}
\title{Setter method for pgxCallableAlleles slot of PGx object}
\usage{
\S4method{pgxCallableAlleles}{PGx}(x) <- value
}
\description{
Setter method for pgxCallableAlleles slot of PGx object
}
