% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{pgxCallableAlleles,PGx-method}
\alias{pgxCallableAlleles,PGx-method}
\title{Getter method for pgxCallableAlleles slot of a PGx object}
\usage{
\S4method{pgxCallableAlleles}{PGx}(x)
}
\value{
Character vector of alleles that can be called
}
\description{
The pgxCallableAlleles slot contains a character vector of allele names for all
allele definitions that are completely represented in the sample VCF and are
thus deemed 'callable'.
}
