% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{pgxGene,PGx-method}
\alias{pgxGene,PGx-method}
\title{Getter method for the pgxGene slot of a PGx object}
\usage{
\S4method{pgxGene}{PGx}(x)
}
\value{
Name of the PGx gene
}
\description{
The 'pgxGene' slot of a PGx object contains the gene name for which the PGx
locations were extracted from the input VCF.
}
