% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-accessors.R
\name{pgxGeneRanges}
\alias{pgxGeneRanges}
\title{Return a GRanges object of the unique ranges for the given gene}
\usage{
pgxGeneRanges(gene, build = "GRCh38")
}
\arguments{
\item{gene}{Gene name}

\item{build}{Genome build. One of "GRCh38" or "GRCh37"}
}
\value{
GRanges object with unique ranges for all haplotypes (star alleles)
for the desired gene
}
\description{
Return a GRanges object of the unique ranges for the given gene
}
\examples{
pgxGeneRanges("CYP2C19")
}
