% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-accessors.R
\name{pgxGenes}
\alias{pgxGenes}
\title{Return all available PGx genes}
\usage{
pgxGenes(build = "GRCh38")
}
\arguments{
\item{build}{Genome build. One of "GRCh38" or "GRCh37"}
}
\value{
character vector of available gene names
}
\description{
Return all available PGx genes
}
\examples{
pgxGenes()
}
