% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{pgxGenotypeCodesToNucleotides,PGx-method}
\alias{pgxGenotypeCodesToNucleotides,PGx-method}
\title{Converts the 'GT' genotype code in a PGx object to nucleotides}
\usage{
\S4method{pgxGenotypeCodesToNucleotides}{PGx}(x, allele, ...)
}
\value{
subsetted PGx object with geno(PGx)$GT converted to nucleotides
}
\description{
Modified from: #' https://github.com/Bioconductor/VariantAnnotation/blob/54d2de6c7a1aa76f404076db79077cd05114edf2/R/methods-VCF-class.R
This method differs in that it uses the REF and ALT columns from the PGx
haplotype definitions instead of the ALT and REF from the sample VCF and
requires that a specific callable allele be used. This method returns a
PGx object that is subsetted by the defined allele, i.e. the input PGx object
is not the same dimensions as the returned PGx object
}
