% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{pgxGenotypeMatrix,PGx-method}
\alias{pgxGenotypeMatrix,PGx-method}
\title{Getter method for pgxGenotypeMatrix}
\usage{
\S4method{pgxGenotypeMatrix}{PGx}(x)
}
\value{
position by sample matrix containing genotypes for all samples in the
PGx object
}
\description{
This method is simply a wrapper around \code{geno(p)$GT}
}
