% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-accessors.R
\name{pgxHaplotypeDefinitions}
\alias{pgxHaplotypeDefinitions}
\title{Return a data.table of the star allele definitions for the given gene}
\usage{
pgxHaplotypeDefinitions(gene, build = "GRCh38")
}
\arguments{
\item{gene}{Gene name}

\item{build}{Genome build. One of "GRCh38" or "GRCh37"}
}
\value{
data.table containing the allele definition matrix for the given gene
}
\description{
Return a data.table of the star allele definitions for the given gene
}
\examples{
pgxHaplotypeDefinitions("CYP2C19")
}
