% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-accessors.R
\name{pgxHaplotypeRanges}
\alias{pgxHaplotypeRanges}
\title{Return a GRanges object of the unique ranges for the given haplotype}
\usage{
pgxHaplotypeRanges(haplotype, build = "GRCh38")
}
\arguments{
\item{haplotype}{Haplotype (star allele) name}

\item{build}{Genome build. One of "GRCh38" or "GRCh37"}
}
\value{
\code{VRanges} object with unique ranges for the given haplotype (star allele)
}
\description{
Return a GRanges object of the unique ranges for the given haplotype
}
\examples{
pgxHaplotypeRanges("CYP2C19_2")
}
