% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-accessors.R
\name{pgxHaplotypes}
\alias{pgxHaplotypes}
\title{Return all available PGx haplotypes (star alleles)}
\usage{
pgxHaplotypes(build = "GRCh38")
}
\arguments{
\item{build}{Genome build. One of "GRCh38" or "GRCh37"}
}
\value{
character vector of available defined star alleles (haplotypes)
}
\description{
Return all available PGx haplotypes (star alleles)
}
\examples{
pgxHaplotypes()
}
