% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{pgxReferenceDataframe<-,PGx-method}
\alias{pgxReferenceDataframe<-,PGx-method}
\title{Setter method for pgxReferenceDataframe slot of PGx object}
\usage{
\S4method{pgxReferenceDataframe}{PGx}(x) <- value
}
\description{
Setter method for pgxReferenceDataframe slot of PGx object
}
