% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{pgxReferenceDataFrame,PGx-method}
\alias{pgxReferenceDataFrame,PGx-method}
\title{Getter method for pgxReferenceDataFrame slot of a PGx object}
\usage{
\S4method{pgxReferenceDataFrame}{PGx}(x)
}
\value{
data.frame of positions by callable haplotypes
}
\description{
The pgxReferenceDataFrame contains the expanded reference strings for every
callable allele of the PGx object
}
