% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-pgx.R
\name{readPGx}
\alias{readPGx}
\title{Read in PGx Data from VCF File}
\usage{
readPGx(file, gene, build = "GRCh38")
}
\arguments{
\item{file}{A phased and normalized VCF file. This file must be indexed (.tbi)}

\item{gene}{The PGx gene to subset from VCF file. See \code{\link{pgxGenes()}} for available genes.}

\item{build}{The genome build. One of "GRCh38" or "GRCh37".}
}
\value{
Object of class PGx
}
\description{
This function is a wrapper around \link[VariantAnnotation]{readVcf()} that
reads only the PGx locations into a PGx object. The user must provide an
indexed VCF file. The path to the indexed VCF file gets passed to
\link[Rsamtools]{TabixFile()} which is then used as the file argument to
\code{readVcf()}.
}
\details{
The PGx locations used to subset the input VCF file are derived from
PharmVar VCF files available at \url{https://www.pharmvar.org/download}.
The "create-reference.R" file included in the data-raw directory of
this package describes the steps used to reproduce the construction of the
GRanges object used in this function. These GRanges objects are accessible
via the \code{pgxAlleleRanges()} function.
}
