% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{scoreHaplotype}
\alias{scoreHaplotype}
\title{Score the haplotype call based on the hamming distance value}
\usage{
scoreHaplotype(x, exact = TRUE)
}
\arguments{
\item{x}{vector of hamming distance values}

\item{exact}{logical. Should only 0-distance (exact matches) be returned?}
}
\description{
Score the haplotype call based on the hamming distance value
}
