% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{scoreMatches}
\alias{scoreMatches}
\title{Score the haplotype call based on the hamming distance value}
\usage{
scoreMatches(x, mismatches = 0)
}
\arguments{
\item{x}{vector of hamming distance values}

\item{mismatches}{logical. Return matches with a score less than or equal to this value}
}
\description{
Score the haplotype call based on the hamming distance value
}
