% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{splitID}
\alias{splitID}
\title{Split the id field of a genotype data.table}
\usage{
splitID(x)
}
\arguments{
\item{x}{data.table}
}
\description{
Split the id field of a genotype data.table
}
