% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-diplotype-matcher.R
\name{subsetCallablePGx}
\alias{subsetCallablePGx}
\title{Subset the PGx object to contain only the positions in the reference allele}
\usage{
subsetCallablePGx(pgx, ref)
}
\arguments{
\item{pgx}{PGx object}

\item{ref}{VRanges object of callable ranges}
}
\value{
list of PGx objects that contain only the range data from each 
callable allele
}
\description{
Subset the PGx object to contain only the positions in the reference allele
}
