% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{summarizeAlleles}
\alias{summarizeAlleles}
\title{Strip suballele ids and return only unique major alleles}
\usage{
summarizeAlleles(x)
}
\arguments{
\item{x}{string of suballele calls}
}
\value{
character string with only unique major star alleles
}
\description{
Strip suballele ids and return only unique major alleles
}
