% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{availableGeneRanges}
\alias{availableGeneRanges}
\title{Return a VRanges object of the unique ranges for the given gene}
\usage{
availableGeneRanges(gene, build = "GRCh38")
}
\arguments{
\item{gene}{Gene name}

\item{build}{Genome build. One of "GRCh38" or "GRCh37"}
}
\value{
\code{VRanges} object with unique ranges for all haplotypes (star alleles)
for the desired gene
}
\description{
Return a VRanges object containing all unique ranges for a given PGx gene 
definition.
}
\examples{
pgxGeneRanges("CYP2C19")
}
