% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-diplotype-matcher.R
\name{callPhasedDiplotypes}
\alias{callPhasedDiplotypes}
\title{Call phased diplotypes}
\usage{
callPhasedDiplotypes(pgx)
}
\arguments{
\item{pgx}{PGx object}

\item{verbose}{Display verbose output? Default TRUE.}
}
\value{
DataFrame of phased allele calls for each sample in PGx
}
\description{
Call phased diplotypes
}
