% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{callPhasedDiplotypes}
\alias{callPhasedDiplotypes}
\title{Call phased diplotypes (star alleles) for all samples in a PGx object}
\usage{
callPhasedDiplotypes(x, mismatches = 0, summarize = TRUE)
}
\arguments{
\item{x}{PGx object}

\item{mismatches}{logical. Report alleles with a hamming distance (number of)
mismatches) less than or equal to this value. Default == 0 (perfect match).}

\item{summarize}{logical. Should star allele calls be summarized up to the
major allele level. Default TRUE. If FALSE, sub-alleles are returned for each
called haplotype.}
}
\value{
data.table with allele calls for each sample in the PGx object
}
\description{
This function attempts to call star allele diplotypes based on the phased
genotype data extracted from the PGx object. It first checks the sample VCF
for any star allele definition that is completely represented. After
determining 'callable' star alleles, the genotype call for each sample is
extracted from the sample VCF and converted to a nucleotide string for all
positions that are both present in the sample VCF and every callable star
allele. Star allele calls are then made by calculating the Hamming distance
between the observed string of phased haplotypes for each sample and each
star allele definition. By default, only star allele definitions that exactly
match (i.e. Hamming distance == 0) are returned. Finally, star allele
calls are summarized to the major allele level.
}
