% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased-allele-caller.R
\name{convertGTtoNucleotides}
\alias{convertGTtoNucleotides}
\title{Convert the GT column of the cleaned data.table to nucleotide calls}
\usage{
convertGTtoNucleotides(x)
}
\arguments{
\item{x}{data.table after cleaning}
}
\description{
Convert the GT column of the cleaned data.table to nucleotide calls
}
