% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{pgxReferenceDataFrame<-,PGx-method}
\alias{pgxReferenceDataFrame<-,PGx-method}
\title{Setter method for pgxReferenceDataFrame slot of PGx object}
\usage{
\S4method{pgxReferenceDataFrame}{PGx}(x) <- value
}
\description{
Setter method for pgxReferenceDataFrame slot of PGx object
}
