% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGx-methods.R
\name{pgxReferenceDataframe,PGx-method}
\alias{pgxReferenceDataframe,PGx-method}
\title{Getter method for pgxReferenceDataframe slot of a PGx object}
\usage{
\S4method{pgxReferenceDataframe}{PGx}(x)
}
\value{
data.frame of positions by callable haplotypes
}
\description{
The pgxReferenceDataframe contains the expanded reference strings for every
callable allele of the PGx object
}
