% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-pgx.R
\name{readPGx}
\alias{readPGx}
\title{Read in PGx Data from VCF File}
\usage{
readPGx(file, gene, build = c("GRCh38", "GRCh37"))
}
\arguments{
\item{file}{A phased and normalized VCF file. This file must be indexed (.tbi)}

\item{gene}{The PGx gene to subset from VCF file. See \code{\link{availableGenes()}} for available genes.}

\item{build}{The genome build. One of "GRCh38" or "GRCh37".}
}
\value{
Object of class PGx
}
\description{
This function is a wrapper around \link[VariantAnnotation]{readVcf()} that
reads only the PGx locations into a PGx object. The user must provide an
indexed VCF file. The path to the indexed VCF file gets passed to
\link[Rsamtools]{TabixFile()} which is then used as the file argument to
\code{readVcf()}.
}
\details{
The PGx locations used to subset the input VCF file are derived from
PharmVar VCF files available at \url{https://www.pharmvar.org/download}.
The "create-reference.R" file included in the data-raw directory of
this package describes the steps used to reproduce the construction of the
GRanges object used in this function. These GRanges objects are accessible
via the \code{availableHaplotypeRanges()} function. The \code{readPGx()} 
function will attempt to detect the type of seqLevels (chromosome names) used 
in the sample VCF (e.g. "UCSC" type - chr1, chr2, etc. or "NCBI" type, 1, 2, 
3, etc.) and modify the ranges of the PharmVar reference accordingly. If the
VCF header does not define a reference then the function will attempt to use
"NCBI" type seqlevels for GRCh37 build and "UCSC" type for GRCh38 builds 
specified by the build argument. If multiple seqLevels are found in the VCF
header then the function will attempt to use the first defined level. 
seqLevels of the final PGx object are set to "UCSC" style for GRCh38 and 
"NCBI" style for GRCh37 to ensure compatibility with downstream processing.
}
