% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarizeDiplotypeCalls}
\alias{summarizeDiplotypeCalls}
\title{Summarize a diplotype call DataFrame}
\usage{
summarizeDiplotypeCalls(df, phased = TRUE)
}
\arguments{
\item{df}{DataFrame or data.frame of diplotype calls with samples as row.names 
and each column corresponding to a star allele}

\item{phased}{logical. Default TRUE. Should the output delimiter indicate that 
the data was phased, i.e. should the output delimiter be "|". If FALSE the 
output delimter is "/"}
}
\value{
DataFrame with summarized calls for each sample
}
\description{
Summarize a diplotype call DataFrame
}
